requirements.txt
absl-py==0.10.0
astroid @ file:///C:/ci/astroid_1592487315634/work
astunparse==1.6.3
blinker==1.4
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cryptography==2.9.2
cvlib==0.2.5
cycler==0.10.0
dnspython==2.0.0
docopt==0.6.2
dominate==2.5.2
email-validator==1.1.1
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.4.4
Flask-Uploads==0.2.1
Flask-WTF==0.14.3
gast==0.3.3
google-auth==1.21.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
imageio==2.9.0
imutils==0.5.3
isort @ file:///C:/ci/isort_1598358138880/work
itsdangerous==1.1.0
Jinja2==2.11.2
jwt==1.0.0
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.3.1
mccabe==0.6.1
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
numpy @ file:///C:/ci/numpy_and_numpy_base_1596215850360/work
oauthlib==3.1.0
olefile==0.46
opencv-python @ file:///C:/Users/Voyager/Downloads/opencv_python-4.4.0.42-cp38-cp38-win_amd64.whl
opt-einsum==3.3.0
Pillow @ file:///C:/ci/pillow_1594298230227/work
pipreqs==0.4.10
progressbar==2.5
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pylint @ file:///C:/ci/pylint_1598617153160/work
PyMySQL==0.10.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.14.0
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.4.1
six==1.15.0
SQLAlchemy==1.3.18
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow-cpu @ file:///C:/Users/Voyager/Downloads/tensorflow_cpu-2.3.0-cp38-cp38-win_amd64.whl
tensorflow-estimator==2.3.0
termcolor==1.1.0
toml @ file:///tmp/build/80754af9/toml_1592853716807/work
torch==1.6.0
torchvision==0.7.0
urllib3==1.25.10
visitor==0.1.3
Werkzeug==1.0.1
wincertstore==0.2
wrapt==1.11.2
WTForms==2.3.3
yarg==0.1.9
